style(features): fix dir setup
