--- src/pages/popup/components/SavedList.tsx
@@ -1,5 +1,0 @@
-const SavedList = () => {
-    return (<p>SavedList</p>)
-}
-
-export default SavedList
