--- src/pages/popup/components/index.ts
@@ -1,3 +1,0 @@
-export { default as Header } from './Header'
-export { default as Input } from './Input'
-export { default as SavedList } from './SavedList'
