+++ src/pages/popup/features/header/Header.tsx
@@ -1,0 +1,13 @@
+import { Stack, Typography } from '@mui/material';
+
+const Header = () => {
+  return (
+    <Stack direction="row" justifyContent="space-between" alignItems="center">
+      <Typography variant="h5" color="black">
+        Character Counter
+      </Typography>
+    </Stack>
+  );
+};
+
+export default Header;
