+++ src/pages/popup/features/index.ts
@@ -1,0 +1,3 @@
+export { default as Header } from './header/Header';
+export { default as Input } from './input/Input';
+export { default as SavedList } from './savedList/SavedList';
