+++ src/pages/popup/features/input/useByteCount.ts
@@ -1,7 +1,7 @@
 import { useMemo } from 'react';
 
 const useByteCount = (inputValue: string) => {
-	return useMemo(() => new Blob([inputValue]).size, [inputValue]);
+  return useMemo(() => new Blob([inputValue]).size, [inputValue]);
 };
 
 export default useByteCount;
