+++ src/pages/popup/features/input/useCharacterCount.ts
@@ -2,6 +2,6 @@ import { useMemo } from 'react';
 
 const useCharacterCount = (inputValue: string) => {
   return useMemo(() => Array.from(inputValue).length, [inputValue]);
-}
+};
 
-export default useCharacterCount
+export default useCharacterCount;
