+++ src/pages/popup/features/savedList/SavedList.tsx
@@ -1,0 +1,5 @@
+const SavedList = () => {
+  return <p>SavedList</p>;
+};
+
+export default SavedList;
