--- src/pages/popup/hooks/index.ts
@@ -1,2 +1,0 @@
-export { default as useCharacterCount } from './useCharacterCount'
-export { default as useByteCount } from './useByteCount'
